### 2021-07-19 #3511 @whedon
[PRE REVIEW]: DBSP_DRP: A Python package for automated spectroscopic data reduction of DBSP data
**Submitting author:** @finagle29 (<a href="http://orcid.org/0000-0003-1118-3132">Milan Roberson</a>)
**Repository:** <a href="https://github.com/finagle29/DBSP_DRP/" target ="_blank">https://github.com/finagle29/DBSP_DRP/</a>
**Version:** v0.9.0
