### 2021-07-21 37fe5c5
make splicing.aco handle no wavelength overlap
