### 2021-07-21 ea4721f
fix ql fluxing and ql median combine multiple frames
