### 2021-07-21 e276e6d
ql: show_1dspec -> show_spectrum
