### 2021-07-21 4444de1
reuse sensible ylim in adjust_splicing
