### 2021-07-21 3a6fc63
add dbsp_trim script
