+++ setup.cfg
@@ -34,6 +34,7 @@ scripts=
     bin/dbsp_show
     bin/dbsp_adjust_splicing
     bin/dbsp_splice
+    bin/dbsp_trim
 
 [options.package_data]
 * = data/*
