### 2021-07-21 9d04b5b
use entry_points instead of scripts
