+++ setup.cfg
@@ -28,16 +28,18 @@ install_requires=
     yattag
     pypeit
     tqdm
-scripts=
-    bin/dbsp_reduce
-    bin/dbsp_ql
-    bin/dbsp_show
-    bin/dbsp_adjust_splicing
-    bin/dbsp_splice
-    bin/dbsp_trim
 
 [options.package_data]
 * = data/*
+
+[options.entry_points]
+console_scripts =
+    dbsp_reduce = dbsp_drp.p200_redux:entrypoint
+    dbsp_ql = dbsp_drp.quicklook:entrypoint
+    dbsp_show = dbsp_drp.show_spectrum:entrypoint
+    dbsp_adjust_splicing = dbsp_drp.adjust_splicing:entrypoint
+    dbsp_splice = dbsp_drp.manual_splice:entrypoint
+    dbsp_trim = dbsp_drp.trim:entrypoint
 
 [tool:pytest]
 testpaths = "dbsp_drp/tests"
