### 2021-07-22 7ae5982
update typo in paper, remove unused affiliation
