### 2021-08-13 #3511 @arfon
:wave: @olebole, @arjunsavel, @crhea93 – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration here is [_DBSP_DRP: A Python package for automated spectroscopic data reduction of DBSP data_](https://github.com/finagle29/DBSP_DRP/).
