### 2021-08-13 #3612 @whedon
[REVIEW]: DBSP_DRP: A Python package for automated spectroscopic data reduction of DBSP data
**Submitting author:** @finagle29 (<a href="http://orcid.org/0000-0003-1118-3132">Milan Roberson</a>)
**Repository:** https://github.com/finagle29/DBSP_DRP/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
