### 2021-08-18 #3612 @crhea93
This is a very interesting-looking package! While the majority of the documentation is present, I think it would benefit from 3 sections that are currently missing:
 - Functionality API
 - Automated Testing
 - Contributing Guidelines
