### 2021-08-18 ccfcf4b
install dbsp_drp during rtd build
should make dbsp_drp.vrsion available
