+++ .readthedocs.yml
@@ -18,3 +18,8 @@ python:
   version: 3.8
   install:
     - requirements: docs/requirements.txt
+    - method: pip
+      path: .
+
+conda:
+  environment: environment.yml
