### 2021-08-18 0ad4481
fail rtd build on warning to help debug
