### 2021-08-19 finagle29/DBSP_DRP#13 @arjunsavel
JOSS paper edits
This project's JOSS paper is in great shape! I have a few small comments that could improve it further:

- [x] I'd consider swapping bits of the first two paragraphs. It might be better to take the last sentence of the first graf ("This process of converting...") and put it directly after the sentence introducing the 1D spectrum. Afterward, the details of spectroscopy (which are very clearly described!) would follow more neatly as processes that this pipeline *does*. As it currently reads, I think the first graf gets a bit too quickly into the jargon, and a more straightforward lead-in to "this is our pipeline" could be nice.
- [x] Please cite IRAF on first reference.   ([here](https://ui.adsabs.harvard.edu/abs/1986SPIE..627..733T/abstract) [are](https://ui.adsabs.harvard.edu/abs/1993ASPC...52..173T/abstract) links to IRAF citation suggestions [from AAS Journals](https://journals.aas.org/news/software-citation-suggestions/))
- [x] "To measure the spectrum” —> “To measure a spectrum” or “To measure the spectrum of a source”
- [x] Kind of a nitpick, but maybe not “constituent wavelengths” — there’s not an intrinsic set of wavelengths that light is split into. Perhaps “wavelength bins” or something similar?
- [x] “, and can open a GUI” —> “, and it can open a GUI”
- [x] “instrument building” —> “instrument-building”
- [x] “The pipeline…the instrument” —> “A pipeline…an instrument”
- [x] “35 year old” —> “35-year-old”
- [x] Parentheses around “e.g. Ogaz & Tollerud (2018)”
- [x] I’d say “sans existing pipeline” instead of just “sans pipeline” — the user would just have to make their own!
- [x] In addition to differences between current and previous observations, nights with highly variable cloud cover might lead to further uncertainty in the number of received photons. Might be worth mentioning, but I’m not very tied to this. Furthermore, the object could have have high / uncertain proper motion (also not tied to this)
- [x] “, that can do a rudimentary” — “, which can do a rudimentary”

This issue is part of my [JOSS review.](https://github.com/openjournals/joss-reviews/issues/3612)
