### 2021-08-19 finagle29/DBSP_DRP#14 @arjunsavel
Documentation edits
This project is nicely documented — I particularly appreciated all the installation details! I have a few comments overall:

- [x] There doesn't appear to be a statement of need in the documentation. I'd recommend adding a new page to this effect or stating this on the [landing page](https://dbsp-drp.readthedocs.io/en/v0.9.0/index.html).
- [x] The API documentation doesn't appear to be working — while the modules are listed, the actual functions/methods and their docstrings are not listed. I'd recommend something like [autodoc](https://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html) if you aren't already using something similar!
- [ ] docstrings are not available for all functions. e.g., manual_tracing.py.ManualTracingGUI.key_press_callback. For acceptance, confirming that there exist docstrings for at least all core functionality is fine!
- [x] I believe that the Docker containers mentioned in the paper aren’t present in the readthedocs.

I also have a few specific suggestions:

- [x] *Installing DBSP_DRP*
  - [x] pacakage —> package
  - [x] What is the expected result of run_pypeit -h?
- [x] *Using the Manual Tracing and Manual Aperture GUIs*
  - [x] matplotlib —> Matplotlib
  - [x] previously-marked —> previously marked
  - [x] and press d to delete —> and pressing d to delete
- [x] *DBSP_DRP QA Outputs*
  - [x] maybe define QA? I see that the linked Pypelt QA page defines it, but might as well put it here, too.
  - [x] Period after "here"
- [x] *Data Reduction Outputs*
  - [x] Colon after find
  - [x] 1d —> 1D
  - [x] one letter —> one-letter
  - [x] numpy —> NumPy
  - [x] astropy —> Astropy
  - [x] fits file —> FITS file

This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/3612).
