### 2021-08-19 finagle29/DBSP_DRP#15 @arjunsavel
Add community guidelines
This could be added either as a page in the documentation or as an extra file in the root of this repository.

This issue is part of my [JOSS review.](https://github.com/openjournals/joss-reviews/issues/3612)
