### 2021-08-19 finagle29/DBSP_DRP#16 @arjunsavel
Update testing instructions
I was able to execute the tests for this project without difficulty — nicely done! I did have a few broad testing-related comments:

- [ ]  Lots of warnings (1671) were generated for me while running the tests. Many were the same (VerifyWarning). If this error can’t be worked around, I’d recommend suppressing it or explicitly mentioning to the user that these are expected in the documentation. Furthermore, single RuntimeWarning was generated in test_quicklook_blue, which should perhaps be investigated. Nitpick for some of the other warnings: np.float —> float, np.bool —> bool, np.int —> int. I also noted a UserWarning about "bad entries." [On a subsequent look, it seems as though many of these warnings are thrown because of PypeIt (I got 478 warnings just from running the PypeIt test suite) — in which case a mention in your documentation should suffice]
- [ ] I'd point people explicitly to running PypeIt's test suite.
- [ ] It might be worth mentioning in the documentation how long you expect the tests to take. On my machine, the tests took ~6 minutes — which is fine, but if a user is expecting a snappy runtime, they might worry that something's wrong!

This issue is part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/3612).
