### 2021-08-19 #3612 @arjunsavel
I’ve completed my review!

Overall:
The contributors to DBSP_DRP have produced an extensive and useful data reduction pipeline. The degree of its automation is impressive, and the contributors’ effort is evident — and there is clear scientific merit, with application to a well-known instrument on a well-known telescope. As someone who’s been on observing runs (though I wouldn’t call myself an observer!) and worked on a reduction pipeline, I’m particularly interested in this code's quicklook capability this code has.

I've opened a few issues in the repository. Many of the items I've mentioned are related to grammar and consistency.

- https://github.com/finagle29/DBSP_DRP/issues/16
- https://github.com/finagle29/DBSP_DRP/issues/15
- https://github.com/finagle29/DBSP_DRP/issues/14
- https://github.com/finagle29/DBSP_DRP/issues/13

I also wanted to touch base about the authorship list — I only see the first author in the commit history.

The various installation routes worked well for me. However, I couldn't fully verify the code's functionality — when I tried to reduce the sample_data, the fact that the data didn’t have RA / Dec ended ups throwing an Astropy units type error. Is there different data I should use to verify that this works, or am I missing something?

In summary, this package looks great to me. I’ll sign off for acceptance pending the resolution of the above issues!
