### 2021-08-19 49d0b2a
rm sys.path manipulation from conf.py
