+++ docs/source/conf.py
@@ -12,7 +12,6 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../..'))
 import sphinx_rtd_theme
 
 
