### 2021-08-19 1739a19
update conf.py copyright
