### 2021-08-19 4bad520
add continue-on-error logic to test and test_dev jobs
