### 2021-08-20 3bb7359
[REVERT THIS] make test_dev use old pypeit version, update pypeit req in environment.yml
