+++ environment.yml
@@ -9,7 +9,7 @@ dependencies:
 - scipy >= 1.4
 - matplotlib >= 3.1
 - pyside2 >= 5.12
-- pypeit == 1.4.2
+- pypeit == 1.5
 - pytest >= 3.0.7
 - yattag >= 1.13
 - tqdm >= 4.52
