### 2021-08-20 473be98
fix continue-on-error logic
