### 2021-08-20 67d650a
add version to conf.py
