+++ docs/source/conf.py
@@ -13,6 +13,7 @@
 import os
 import sys
 import sphinx_rtd_theme
+import dbsp_drp
 
 
 # -- Project information -----------------------------------------------------
@@ -21,6 +22,7 @@ project = 'DBSP_DRP'
 copyright = '2020-2021, California Institute of Technology'
 author = 'Milan S. Roberson'
 
+version = dbsp_drp.__version__
 
 # -- General configuration ---------------------------------------------------
 
