+++ dbsp_drp/reduction.py
@@ -27,6 +27,18 @@ from dbsp_drp.manual_tracing import ManualTracingGUI
 
 def parse_pypeit_parameter_file(parameter_file: str,
         spectrograph: str) -> List[str]:
+    """
+    Grab user-provided PypeIt configuration for ``spectrograph`` from
+    ``parameter_file``, which contains user-provided PypeIt configuration for
+    both arms of DBSP.
+
+    Args:
+        parameter_file (str): User-provided file with their PypeIt config.
+        spectrograph (str): PypeIt name of spectrograph.
+
+    Returns:
+        List[str]: User-provided PypeIt configuration for ``spectrograph``
+    """
     user_config_lines = []
     read_lines = False
     arm = 'red' if 'red' in spectrograph else 'blue'
@@ -47,9 +59,9 @@ def parse_pypeit_parameter_file(parameter_file: str,
 
 def search_table_for_arc(row: Row, i: int, table: Table, step: int, max_sep: Angle) -> Tuple[int, int]:
     """
-    Searches table starting at row i in steps of j_init / abs(j) for frames
-    within max_separation. If an arc frame is found, the arc's calib is
-    assigned to row.
+    Searches table starting at row ``i`` in steps of ``step`` for frames within
+    ``max_separation`` of ``row``. If an arc frame is found, the arc's ``calib``
+    is assigned to ``row``.
 
     Args:
         row (Row): starting row
@@ -142,8 +154,16 @@ def set_calibs(table: Table):
 def setup(file_list: List[str], output_path: str, spectrograph: str) -> Tuple[PypeItSetup, str]:
     """
     Does PypeIt setup, without writing the .pypeit file
-    """
 
+    Args:
+        file_list (List[str]): List of raw data files to reduce.
+        output_path (str): reduction output path
+        spectrograph (str): PypeIt name of spectrograph.
+
+    Returns:
+        Tuple[PypeItSetup, str]: PypeItSetup object, reduction output path
+    """
+    # TODO: do we need to check if output_path is None or return it?
     # Get the output directory
     output_path = os.getcwd() if output_path is None else output_path
     sort_dir = os.path.join(output_path, 'setup_files')
@@ -174,6 +194,18 @@ def write_setup(context: Tuple[PypeItSetup, str], cfg_split: str,
         spectrograph: str, user_config_lines: List[str]) -> List[str]:
     """
     Writes the .pypeit file
+
+    Args:
+        context (Tuple[PypeItSetup, str]): PypeItSetup object, reduction output path
+        cfg_split (str): [description]
+        spectrograph (str): PypeIt name of spectrograph.
+        user_config_lines (List[str]): User-provided PypeIt configuration.
+
+    Raises:
+        RuntimeError: Raised if files were not assigned a frame type.
+
+    Returns:
+        List[str]: List of PypeIt files generated.
     """
     ps, output_path = context
 
@@ -203,6 +235,16 @@ def redux(pypeit_file: str, output_path: str, reuse_masters: bool = True,
         show: bool = False, calib_only: bool = False) -> Tuple[set, set]:
     """
     Runs the reduction
+
+    Args:
+        pypeit_file (str): Path to PypeIt reduction file.
+        output_path (str): reduction output path
+        reuse_masters (bool, optional): Reuse master calibration files (if they exist). Defaults to True.
+        show (bool, optional): Show debugging/intermediate plots. Defaults to False.
+        calib_only (bool, optional): Only perform calibration? Defaults to False.
+
+    Returns:
+        Tuple[set, set]: set of filenames of reduced (spec1d, spec2d) files.
     """
     splitnm = os.path.splitext(pypeit_file)
     if splitnm[1] != '.pypeit':
[…]
