### 2021-09-21 717edbf
add docs for dbsp_trim, dbsp_show, docker images
