### 2021-09-22 dc1e007
add conda install instructions
