### 2021-09-22 finagle29/DBSP_DRP#14 @finagle29
Thanks for the thorough suggestions for documentation improvements! I've implemented all of them, and can confirm that docstrings exist for all core functionality.
