+++ paper/paper.bib
@@ -102,7 +102,7 @@ archivePrefix = {arXiv},
 @ARTICLE{De2020,
        author = {{De}, Kishalay and {Kasliwal}, Mansi M. and {Tzanidakis}, Anastasios and {Fremling}, U. Christoffer and {Adams}, Scott and {Aloisi}, Robert and {Andreoni}, Igor and {Bagdasaryan}, Ashot and {Bellm}, Eric C. and {Bildsten}, Lars and {Cannella}, Christopher and {Cook}, David O. and {Delacroix}, Alexandre and {Drake}, Andrew and {Duev}, Dmitry and {Dugas}, Alison and {Frederick}, Sara and {Gal-Yam}, Avishay and {Goldstein}, Daniel and {Golkhou}, V. Zach and {Graham}, Matthew J. and {Hale}, David and {Hankins}, Matthew and {Helou}, George and {Ho}, Anna Y.~Q. and {Irani}, Ido and {Jencson}, Jacob E. and {Kaplan}, David L. and {Kaye}, Stephen and {Kulkarni}, S.~R. and {Kupfer}, Thomas and {Laher}, Russ R. and {Leadbeater}, Robin and {Lunnan}, Ragnhild and {Masci}, Frank J. and {Miller}, Adam A. and {Neill}, James D. and {Ofek}, Eran O. and {Perley}, Daniel A. and {Polin}, Abigail and {Prince}, Thomas A. and {Quataert}, Eliot and {Reiley}, Dan and {Riddle}, Reed L. and {Rusholme}, Ben and {Sharma}, Yashvi and {Shupe}, David L. and {Sollerman}, Jesper and {Tartaglia}, Leonardo and {Walters}, Richard and {Yan}, Lin and {Yao}, Yuhan},
         title = "{The Zwicky Transient Facility Census of the Local Universe. I. Systematic Search for Calcium-rich Gap Transients Reveals Three Related Spectroscopic Subclasses}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Supernovae, Compact objects, White dwarf stars, 1668, 288, 1799, Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics},
          year = 2020,
         month = dec,
@@ -115,5 +115,109 @@ archivePrefix = {arXiv},
        eprint = {2004.09029},
  primaryClass = {astro-ph.HE},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...905...58D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@INPROCEEDINGS{IRAF86,
+       author = {{Tody}, Doug},
+        title = "{The IRAF Data Reduction and Analysis System}",
+    booktitle = {Instrumentation in astronomy VI},
+         year = 1986,
+       editor = {{Crawford}, David L.},
+       series = {Society of Photo-Optical Instrumentation Engineers (SPIE) Conference Series},
+       volume = {627},
+        month = jan,
+        pages = {733},
+          doi = {10.1117/12.968154},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1986SPIE..627..733T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@INPROCEEDINGS{IRAF93,
+       author = {{Tody}, Doug},
+        title = "{IRAF in the Nineties}",
+    booktitle = {Astronomical Data Analysis Software and Systems II},
[…]
