### 2021-10-05 #3612 @finagle29
Thanks for checking in!

I've been slowly working through the review issues that @arjunsavel opened. I've been a bit stuck on how to write contributing guidelines. Regarding the authorship list, Christoffer Fremling and Mansi Kasliwal were both substantial non-code contributors. The sample_data folder is slightly misnamed, it's intentionally broken FITS files that the unit tests rely on. I have put sample data to verify functionality here: https://drive.google.com/drive/folders/1HcbmItr50IN2D4GF37ycoZHa1Cco__6m?usp=sharing

I have two quick clarifying question for @crhea93. By functionality API, do you mean something like this automatically generated API documentation https://dbsp-drp.readthedocs.io/en/develop/dbsp_drp.html ? Also, could you elaborate on what Automated Testing documentation you would like to see?
