### 2021-10-06 797c832
explicitly create HIERARCH cards
this avoids raising astropy.io.fits.verify.VerifyWarnings when creating new header keywords that are longer than 8 characters.
