### 2021-10-06 8f5aba5
docs: add warnings/runtime to post-installation testing
