### 2021-10-06 1539eae
shuffle pytest config to pyproject.toml
