+++ pyproject.toml
@@ -5,3 +5,6 @@ requires = [
     "setuptools_scm[toml]>=3.4"
 ]
 build-backend = 'setuptools.build_meta'
+
+[tool.pytest.ini_options]
+testpaths = "dbsp_drp/tests"
