### 2021-10-06 63176d4
paper: add outer parens to (Ogaz & Tollerud (2018))
