### 2021-10-06 f753b2b
clarify dir name for table_edit test files
