### 2021-10-06 finagle29/DBSP_DRP#13 @finagle29
I think with #21 merged, I've addressed all of these.
