### 2021-10-06 04b68ac
first pass on community guidelines
