+++ docs/source/community.rst
@@ -1,0 +1,44 @@
+===================
+Community Guidlines
+===================
+
+Contributing
+============
+Contributions come in many different kinds, spanning bug fixes and new features
+to suggestions for improved documentation or ways to make your DBSP_DRP
+workflow easier for you and more efficient. All types of contributions from the
+community are greatly welcomed and much appreciated.
+
+If you want to fix a bug or contribute a new feature, please create a fork of
+DBSP_DRP on GitHub, make changes to your fork, and then create a Pull Request
+when your changes are ready to be reviewed for inclusion in DBSP_DRP.
+Maintainers will review your proposed changes and give you feedback. Once all
+parties are satisfied with the proposed changes, your code will be merged.
+Please create pull requests to the develop branch of DBSP_DRP, as this is where
+active development occurs.
+
+If you have suggestions for how to improve DBSP_DRP, please open an Issue on
+GitHub describing how you want DBSP_DRP to be improved. Maintainers and
+developers will comment on the Issue, discussing your suggestions in a
+collaborative manner, developing and maturing it with you until it is ready to
+be implemented (and someone has volunteered to do so).
+
+Reporting Issues
+================
+If you run into an error while running DBSP_DRP, identify a bug, documentation
+deficiency, or potential enhancement, please file an
+`Issue on GitHub <https://github.com/finagle29/DBSP_DRP/issues>`_. For bugs or
+errors, please try to be as descriptive as possible when writing what went
+wrong -- it's hard to provide too much detail! At a bare minimum, please
+include the versions of Python, DBSP_DRP, PypeIt, NumPy, and Astropy that were
+installed when you encountered the error. If you can create a "minimum working
+example" for the error, the smallest set of input files and command line
+invocation or Python code that still manifests this error, and share this with
+your Issue, it will greatly assist anybody who tries to fix your error.
+
+Getting Support
+===============
+If you are having trouble running DBSP_DRP, you can open a Discussion on GitHub
+to ask the community for help. There are no silly questions -- though we try to
+make DBSP_DRP work as well as possible, computers are finnicky creatures and
+entropy is always working against us.
