### 2021-10-10 #3612 @arjunsavel
Hi @finagle29 — thank you for addressing my issues, things are looking great! I've closed all the completed issues.

With regard to community guidelines, you could use the emcee package's [`CONTRIBUTING.md`](https://github.com/dfm/emcee/blob/main/CONTRIBUTING.md) file as an example. In short, you'll need sections describing how to contribute to your package, how to report issues/bugs, and how to get help. Let me know if you have any specific questions about these!

Also, thanks for pointing me to some sample data — I'll be able to check that out in the next few days.
