### 2021-10-26 01f68ff
move present affiliation entirely to author footnote
