### 2021-10-27 7057099
rm debugging code, unused docker compose
