### 2021-10-27 02cdc4b
bugfix: stop crash on arc without ra/dec
