### 2021-10-27 5f2a09f
update readme: citation and install
