+++ README.md
@@ -22,14 +22,22 @@ It adds several GUIs to allow for easier control of your reduction:
 The latest documentation can be found on [Read the Docs](https://dbsp-drp.readthedocs.io/en/latest/index.html).
 
 ## Citation
-If you use DBSP_DRP in your research, please cite this repository
-
-`DBSP_DRP (2021). https://github.com/finagle29/DBSP_DRP`
-
-or use this BibTeX reference
+If you use DBSP_DRP in your research, please cite the following publication,
+and optionally this repository itself.
 
 ```
-@misc{dbsp_drp2021,
+@misc{dbsp_drp2021_arxiv,
+      title={DBSP_DRP: A Python package for automated spectroscopic data reduction of DBSP data}, 
+      author={Milan S. Roberson and Christoffer Fremling and Mansi M. Kasliwal},
+      year={2021},
+      eprint={2107.12339},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.IM}
+}
+```
+
+```
+@misc{dbsp_drp2021_github,
     title     = "DBSP_DRP",
     year      = "2021",
     publisher = "GitHub",
@@ -41,10 +49,11 @@ and please also cite [PypeIt](https://github.com/pypeit/PypeIt#citation).
 
 ## Prerequisites
 DBSP_DRP's dependencies are detailed in [environment.yml](environment.yml).
-You can install all prerequisites by downloading the environment.yml file, navigating to the directory containing it in your terminal window and running
+You can install all prerequisites for a `pip` or source install by downloading the environment.yml file, navigating to the directory containing it in your terminal window and running
 ```shell_session
 $ conda env create -f environment.yml
 ```
+Installing DBSP_DRP using `conda` does not require this step.
 
 The telluric correction code provided by PypeIt relies on a large (5 GB) atmospheric model file,
 TellFits_Lick_3100_11100_R10000.fits, which can be downloaded
@@ -54,7 +63,12 @@ and must be installed into the ``pypeit/data/telluric/`` directory of your PypeI
 An easier alternative is to use the [download_tellfile](bin/download_tellfile) script to download and install the atmospheric model file for you.
 
 ## Installation
-You can install using `pip`
+You can install using `conda`
+```shell_session
+$ conda install -c conda-forge dbsp_drp
+```
+
+or `pip`
 ```shell_session
 $ pip install dbsp-drp
 ```
