### 2021-10-27 b680a75
update test badge to something more meaningful
