### 2021-11-05 #3612 @arjunsavel
Hi @finagle29 — sorry for the delay!

I've been able to run the sample_data that you provided for the blue arm, and everything looks great there! However, I'm not able to do so for the red arm; in this case, I run into a consistent issue with each file (no RA, Dec --> astropy throws a TypeError when trying to create SkyCoords).
