### 2021-11-09 2054fab
check for missing RA/DEC instead of using airmass as proxy
since None ra/dec are the underlying cause of crashes this bugfix is meant to address, let's explicitly check for ra/dec instead of using airmass as a proxy
