### 2021-11-09 #3612 @finagle29
No worries, @arjunsavel !

I ran into that bug myself, and have just pushed a fix to the `develop` branch. If this error (or other errors) persists, we can take discussion over to https://github.com/finagle29/DBSP_DRP/discussions.
