### 2021-11-13 #3612 @finagle29
> I have two quick clarifying question for @crhea93. By functionality API, do you mean something like this automatically generated API documentation https://dbsp-drp.readthedocs.io/en/develop/dbsp_drp.html ? Also, could you elaborate on what Automated Testing documentation you would like to see?

Hi @crhea93, I just wanted to follow up on these questions I had regarding your comments
