### 2022-01-10 9dfca45
adjust_splicing: check that both red and blue data are present
