### 2022-01-10 ae6b2d5
ql: prevent crash when ra/dec data is misisng
