### 2022-01-10 #3612 @finagle29
> > And a question for @arfon : I have changed institutional affiliations since completing the vast majority of my work on DBSP_DRP, should I use my previous affiliation on the paper? If so, is there some way I should note my current affiliation?
>
> You can include multiple affiliations if you wish? Perhaps listing your current and previous ones? The [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) shows how to do this.

Thanks for the example! I decided on listing my present address/affiliation in a footnote.
