### 2022-01-11 #3612 @arjunsavel
Happy new year! @arfon, I've just completed going over some bug fixes that @finagle made — everything's looking great, and I'd consider my review completed.
