### 2022-02-23 #3612 @finagle29
Hi @arfon, I've created a new release including the changes from this review and archived it in Zenodo. The URL that points to this specific version (version 1.0.0) is https://doi.org/10.5281/zenodo.6241526, and https://doi.org/10.5281/zenodo.6241525 points to the latest version, I'm unsure which is better to use here.
