### 2022-02-23 fb264c7
hack to fix version displayed on rtd
