+++ setup.py
@@ -1,5 +1,6 @@
 import os
 from setuptools import setup
+import shutil
 
 VERSION_TEMPLATE = """
 # Note that we need to fall back to the hard-coded version if either
@@ -11,6 +12,11 @@ try:
 except Exception:
     version = '{version}'
 """.lstrip()
+
+with open('docs/source/conf.orig.py', 'r') as src:
+    with open('docs/source/conf.py', 'w') as dst:
+        dst.write("# THIS FILE WAS COPIED FROM conf.orig.py. DO NOT MODIFY\n")
+        shutil.copyfileobj(src, dst)
 
 setup(use_scm_version={'write_to': os.path.join('dbsp_drp', 'version.py'),
                        'write_to_template': VERSION_TEMPLATE})
