### 2022-02-28 #3612 @finagle29
I just noticed a small issue: in the published version of the paper, there isn't a footnote after my name noting my present address/affiliation. When I use whedon to preview the paper on the `main` branch of the repo, the footnote is present.
