### 2022-03-01 #3612 @arfon
I wonder if we're stripping this? As [per the docs](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), the preferred way would be to write this as such:

```
authors:
  - name: Milan S. Roberson^[Present address: Department of Physics and Astronomy, University of California, Los Angeles.]
    orcid: 0000-0003-1118-3132
    affiliation: "1,2"
  - name: Christoffer Fremling
    orcid: 0000-0002-4223-103X
    affiliation: 2
  - name: Mansi M. Kasliwal
    orcid: 0000-0002-5619-4938
    affiliation: 2
affiliations:
  - name: Schmidt Academy of Software Engineering, California Institute of Technology
    index: 1
  - name: Division of Physics, Mathematics and Astronomy, California Institute of Technology
    index: 2
date: 26 October 2021
bibliography: paper.bib
nocite: |
  @Lunnan2020
---
```
