### 2022-03-01 #3612 @tarleb
The note should now be rendered in the PDF in the same way it did before. I'm now checking how the note can be included in JATS output.
