### 2022-03-12 #3612 @finagle29
Thanks @tarleb !!
Is it possible to rerender the PDF displayed at https://joss.theoj.org/papers/10.21105/joss.03612 ?
