### 2022-03-21 #3612 @finagle29
@arfon could you rerender the published PDF?
I noticed this just now, but as a result of the footnote issue, the citation string on the JOSS website includes a open square bracket as the last character of my last name:
`Roberson[ et al., (2022). DBSP_DRP: A Python package for automated spectroscopic data reduction of DBSP data. Journal of Open Source Software, 7(70), 3612, https://doi.org/10.21105/joss.03612`
