### 2022-03-22 #3612 @arfon
> @arfon could you rerender the published PDF?

This looks to be fixed now @finagle29?
