### 2022-10-02 #3612 @arfon
@finagle29 – can you check this new PDF looks good? https://github.com/openjournals/joss-papers/blob/13c81ea7f6403838a32bbfbc9ae0835754b5e7d9/joss.03612/10.21105.joss.03612.pdf
